+++ subjects/Hunk-m68k/BYTEOPS.reko/BYTEOPS_code.c
@@ -418,7 +418,7 @@ Eq_n fn00001458(Eq_n d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 							uint32 d0_n = (word16) d2_n *32 0x0A + SEQ(SLICE(__swap(SEQ(SLICE(d1_n, word16, 16), (word16) d1_n + (word16) __swap(d2_n) * 0x0A)), word16, 16), 0x00);
 							++a2_n;
 							a0_n = (struct Eq_n *) ((char *) &g_t275D + (SEQ(SLICE(d0_n, word24, 8), (a2_n->t0000).u0) & 0xFF));
-							int32 d1_n = (int32) (int16) a2_n->t0000.u0;
+							int32 d1_n = (int32) a2_n->t0000.u0;
 							d0_n = (uint32) a0_n->t0000.u1;
 							d1_n = d1_n - 0x30;
 							d2_n = d1_n - 0x30 + d0_n;
@@ -449,7 +449,7 @@ Eq_n fn00001458(Eq_n d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 							uint32 d1_n = __swap((word32 *) 0x0A) *32 (word16) d5_n;
 							uint32 d0_n = (word16) d5_n *32 0x0A + SEQ(SLICE(__swap(SEQ(SLICE(d1_n, word16, 16), (word16) d1_n + (word16) __swap(d5_n) * 0x0A)), word16, 16), 0x00);
 							++a2_n;
-							int32 d1_n = (int32) (int16) a2_n->t0000.u0;
+							int32 d1_n = (int32) a2_n->t0000.u0;
 							a0_n = (struct Eq_n *) ((char *) &g_t275D + (SEQ(SLICE(d0_n, word24, 8), (a2_n->t0000).u0) & 0xFF));
 							d1_n = d1_n - 0x30;
 							d5_n = d1_n - 0x30 + d0_n;
@@ -459,7 +459,7 @@ Eq_n fn00001458(Eq_n d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 				if (a2_n->t0000.u0 == 0x68 || ((a2_n->t0000).u0 == 0x6C || ((a2_n->t0000).u0 == 0x4C || ((a2_n->t0000).u0 == 0x6A || ((a2_n->t0000).u0 == 122 || (a2_n->t0000).u0 == 116)))))
 				{
 					++a2_n;
-					d4_n = (int32) (int16) a2_n->t0000.u0;
+					d4_n = (int32) a2_n->t0000.u0;
 				}
 				if (d4_n == 0x68 && (a2_n->t0000).u0 == 0x68)
 				{
@@ -639,7 +639,7 @@ l000016D6:
 						uint8 bLoc31_n = SLICE(d1_n, byte, 24);
 						uint64 qwLoc38_n = SEQ(dwLoc38_n, d1_n);
 						if (d4_n == 0x02)
-							qwLoc38_n = (uint64) (uint32) bLoc31_n;
+							qwLoc38_n = (uint64) bLoc31_n;
 						dwLoc38_n = (word32) qwLoc38_n;
 						dwLoc34_n = SLICE(qwLoc38_n, word32, 32);
 						bLoc44_n = (byte) d1_n;
@@ -705,7 +705,7 @@ l00001932:
 							if (d4_n == 0x02)
 							{
 								Eq_n d1_n;
-								d1_n.u0 = (int32) (int16) bLoc39_n;
+								d1_n.u0 = (int32) bLoc39_n;
 								d1_n = SEQ(SLICE(d0_n, word24, 8), bLoc39_n);
 								qwLoc40_n = SEQ(d1_n >> 31, d1_n);
 							}
@@ -786,7 +786,7 @@ l00001BC4:
 					{
 						do
 						{
-							if (fn00001DD8((int32) (int16) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+							if (fn00001DD8((int32) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 							{
 								d0_n = d6_n;
 								return d0_n;
@@ -824,7 +824,7 @@ l00001BC4:
 						do
 						{
 							word32 a0_n;
-							if (fn00001DD8((int32) (int16) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+							if (fn00001DD8((int32) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 							{
 								d0_n = d6_n;
 								return d0_n;
@@ -855,7 +855,7 @@ l00001BC4:
 					do
 					{
 						word32 a0_n;
-						if (fn00001DD8((int32) (int16) *(a3_n + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+						if (fn00001DD8((int32) *(a3_n + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 						{
 							d0_n = d6_n;
 							return d0_n;
@@ -888,7 +888,7 @@ l00001BC4:
 			{
 				++a4_n;
 				word32 a0_n;
-				d0 = fn00001DD8((int32) (int16) *a4_n, a5_n, out a0_n, out a5_n);
+				d0 = fn00001DD8((int32) *a4_n, a5_n, out a0_n, out a5_n);
 				if (d0 == ~0x00)
 				{
 					d0_n = d6_n;
