+++ subjects/Hunk-m68k/FIBO.reko/FIBO_code.c
@@ -469,7 +469,7 @@ void fn000014B4(ui32 d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 							uint32 d0_n = (word16) d2_n *32 0x0A + SEQ(SLICE(__swap(SEQ(SLICE(d1_n, word16, 16), (word16) d1_n + (word16) __swap(d2_n) * 0x0A)), word16, 16), 0x00);
 							++a2_n;
 							a0_n = (struct Eq_n *) (g_a28B1 + (SEQ(SLICE(d0_n, word24, 8), (a2_n->t0000).u0) & 0xFF));
-							int32 d1_n = (int32) (int16) a2_n->t0000.u0;
+							int32 d1_n = (int32) a2_n->t0000.u0;
 							d0_n = (uint32) a0_n->t0000.u1;
 							d1_n = d1_n - 0x30;
 							d2_n = d1_n - 0x30 + d0_n;
@@ -501,7 +501,7 @@ void fn000014B4(ui32 d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 							uint32 d1_n = __swap(0x0A) *32 (word16) d5_n;
 							uint32 d0_n = (word16) d5_n *32 0x0A + SEQ(SLICE(__swap(SEQ(SLICE(d1_n, word16, 16), (word16) d1_n + (word16) __swap(d5_n) * 0x0A)), word16, 16), 0x00);
 							++a2_n;
-							int32 d1_n = (int32) (int16) a2_n->t0000.u0;
+							int32 d1_n = (int32) a2_n->t0000.u0;
 							a0_n = (struct Eq_n *) (g_a28B1 + (SEQ(SLICE(d0_n, word24, 8), (a2_n->t0000).u0) & 0xFF));
 							d1_n = d1_n - 0x30;
 							d5_n = d1_n - 0x30 + d0_n;
@@ -511,7 +511,7 @@ void fn000014B4(ui32 d0, struct Eq_n * dwArg04, int8 * dwArg08, Eq_n dwArg0C)
 				if (a2_n->t0000.u0 == 0x68 || ((a2_n->t0000).u0 == 0x6C || ((a2_n->t0000).u0 == 0x4C || ((a2_n->t0000).u0 == 0x6A || ((a2_n->t0000).u0 == 122 || (a2_n->t0000).u0 == 116)))))
 				{
 					++a2_n;
-					d4_n = (int32) (int16) a2_n->t0000.u0;
+					d4_n = (int32) a2_n->t0000.u0;
 				}
 				if (d4_n == 0x68 && (a2_n->t0000).u0 == 0x68)
 				{
@@ -692,7 +692,7 @@ l00001732:
 						uint8 bLoc31_n = SLICE(d1_n, byte, 24);
 						uint64 qwLoc38_n = SEQ(dwLoc38_n, d1_n);
 						if (d4_n == 0x02)
-							qwLoc38_n = (uint64) (uint32) bLoc31_n;
+							qwLoc38_n = (uint64) bLoc31_n;
 						dwLoc38_n = (word32) qwLoc38_n;
 						dwLoc34_n = SLICE(qwLoc38_n, word32, 32);
 						bLoc44_n = (byte) d1_n;
@@ -758,7 +758,7 @@ l0000198E:
 							if (d4_n == 0x02)
 							{
 								Eq_n d1_n;
-								d1_n.u0 = (int32) (int16) bLoc39_n;
+								d1_n.u0 = (int32) bLoc39_n;
 								d1_n = SEQ(SLICE(d0_n, word24, 8), bLoc39_n);
 								qwLoc40_n = SEQ(d1_n >> 31, d1_n);
 							}
@@ -841,7 +841,7 @@ l00001C20:
 					{
 						do
 						{
-							if (fn00001E34((int32) (int16) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+							if (fn00001E34((int32) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 								return;
 							d2_n = (word32 *) ((char *) d2_n + 1);
 							d6_n = (word32) d6_n.u0 + 1;
@@ -873,7 +873,7 @@ l00001C20:
 						do
 						{
 							word32 a0_n;
-							if (fn00001E34((int32) (int16) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+							if (fn00001E34((int32) *(fp->aFFFFFFFA + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 								return;
 							d2_n = (word32 *) ((char *) d2_n + 1);
 							d6_n = (word32) d6_n.u0 + 1;
@@ -897,7 +897,7 @@ l00001C20:
 				while (d2_n < dwLoc0A_n)
 				{
 					word32 a0_n;
-					if (fn00001E34((int32) (int16) *(a3_n + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
+					if (fn00001E34((int32) *(a3_n + d2_n), a5_n, out a0_n, out a5_n) == ~0x00)
 						return;
 					d2_n.u2 = (word32) d2_n + 1;
 					d6_n = (word32) d6_n.u0 + 1;
@@ -923,7 +923,7 @@ l00001C20:
 			{
 				++a4_n;
 				word32 a0_n;
-				d0 = fn00001E34((int32) (int16) *a4_n, a5_n, out a0_n, out a5_n);
+				d0 = fn00001E34((int32) *a4_n, a5_n, out a0_n, out a5_n);
 				if (d0 == ~0x00)
 					return;
 				d6_n = (word32) d6_n.u0 + 1;
@@ -2747,7 +2747,7 @@ l0000392E:
 						Eq_n d1_n;
 						d1 = d1_n;
 					}
-					d0.u2 = (int32) (int16) a4_n->t0000.u0;
+					d0.u2 = (int32) a4_n->t0000.u0;
 					d5_n = dwLoc4C_n;
 					d3_n.u0 = 0x01;
 					d4_n.u2 = (word32) d4_n + 1;
@@ -2903,7 +2903,7 @@ int32 fn00003E04(int8 * dwArg04)
 			{
 				word32 a0_n;
 				word32 a5_n;
-				d1_n = fn00001E90((int32) (int16) *a2_n, *(struct Eq_n **) 0x3FF4, out a0_n, out a5_n);
+				d1_n = fn00001E90((int32) *a2_n, *(struct Eq_n **) 0x3FF4, out a0_n, out a5_n);
 			}
 			if (d1_n == -1)
 			{
