+++ subjects/PE/m68k/hello_m68k_CRTSTDIO.c
@@ -226,7 +226,7 @@ struct Eq_n * fn00001C40(struct Eq_n * a5, byte bArg04, struct Eq_n * dwArg08, b
 	{
 		*dwArg08->ptr0000 = bArg07_n;
 		uint8 * a0_n = dwArg08->ptr0000;
-		d0_n = (int32) (int16) *a0_n;
+		d0_n = (int32) *a0_n;
 		dwArg08->ptr0000 = a0_n + 1;
 	}
 	else
